requirements.txt
gradio>=4.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq
langchain-huggingface
langchain-openai
langchain-chroma
chromadb>=0.4.0
pydantic>=2.0.0
python-dotenv
tiktoken
pandas
tqdm
tenacity
litellm
sentence-transformers